Updated CRUD with authorization

+++ src/main/java/com/cognizant/insurance/dto/AgentDTO.java
@@ -15,6 +15,6 @@ public class AgentDTO {
 	private String email;
 	private String role;
 	
-	private int phno;
+	private Long phno;
 
 }
